# openapi.yml
openapi: 3.0.3
info:
  title: hyper/webcms
  description: >
    ![logo](https://hyper63.com/favicon-96x96.png)

    # Welcome to hyper/webCMS
    
    An content management system for the hyper website
  contact:
    url: https://hyper.io/contact
    email: info@hyper.io
    name: hyper
  license:
    name: Apache Licnense 2.0
    url: https://www.github.com/hyper63/webcms/licence
  version: "1.0"
servers:
  - description: Development Server
    url: http://localhost:3000
  - description: Production Server
    url: https://webcms.hyper.io
  - description: API Production Server
    url: https://webcms-api.hyper.io/
paths:
  /:
    get: 
      tags: [health]
      description: Healthcheck endpoint
      responses:
        "200": 
          description: successful health response
          content:
            application/json:
              schema:
                properties:
                  name:
                    description: webcms
                    type: string
  /api/faqs/{id}:
    get:
      description: Retrieve a faq by id
      tags: [faqs]
      parameters:
        - name: id
          in: path
          description: accepts an faq id value
          required: true
          schema:
            type: string
      responses:
        "200":
          description: returns a faq by id
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                  question:
                    type: string
                  answer: 
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
        "404":
          description: not found
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        "500":
          description: server error
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
    put:
      description: Updates a FAQ Document
      tags: [faq]
      parameters:
        - name: id
          in: path
          description: accepts an faq id value
          required: true
          schema:
            type: string
      requestBody:
        description: faq document
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  type: string
                question:
                  type: string
                answer: 
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      security:
        - bearerAuth: []
      responses:
        "200":
          description: successfully updated faq document
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  id:
                    type: string
        "400":
          description: bad request failure due to invalid faq document
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        "401":
          description: not authorized
          content:
            application/json:
              schema:
                properties:
                  ok: 
                    type: boolean
                  message:
                    type: string
        "409":
          description: document conflict
          content:
            application/json:
              schema:
                properties:
                  ok: 
                    type: boolean
                  message:
                    type: string
        "500":
          description: server error
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
    delete:
      description: Updates a FAQ Document
      tags: [faq]
      parameters:
        - name: id
          in: path
          description: accepts an faq id value
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        "200":
          description: returns a faq by id
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                  question:
                    type: string
                  answer: 
                    type: string
                  tags:
                    type: array
                    items:
                      type: string
        "404":
          description: not found
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        "500":
          description: server error
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
  /api/faqs:
    post:
      description: Create FAQ Document
      tags: [faq]
      requestBody:
        description: faq document
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  type: string
                question:
                  type: string
                answer: 
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      security:
        - bearerAuth: []
      responses:
        "201":
          description: successfully created faq document
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  id:
                    type: string
        "400":
          description: bad request failure due to invalid faq document
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        "401":
          description: not authorized
          content:
            application/json:
              schema:
                properties:
                  ok: 
                    type: boolean
                  message:
                    type: string
        "409":
          description: document conflict
          content:
            application/json:
              schema:
                properties:
                  ok: 
                    type: boolean
                  message:
                    type: string
        "500":
          description: server error
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
    get:
      description: list faqs
      tags: [faq]
      responses:
        "200":
          description: successful list of faqs
          content:
            application/json:
              schema:
                properties:
                  ok:
                    type: boolean
                  faqs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        question:
                          type: string
                        answer:
                          type: string
                        tags:
                          type: array
                          items:
                            type: string
                        created:
                          type: string
                        updated:
                          type: string
        "500":
          description: server error
          content:
            application/json:
              schema:
                properties:
                  ok: 
                    type: boolean
                  message:
                    type: string
                    

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT